convert float to integer before comparing in PageManager.prevViewport()/nextViewport()

+++ android/kreader/app/src/main/java/com/onyx/kreader/host/math/PageManager.java
@@ -339,7 +339,7 @@ public class PageManager {
     }
 
     public boolean nextViewport() {
-        if (viewportRect.bottom >= pagesBoundingRect.bottom) {
+        if ((int)viewportRect.bottom >= (int)pagesBoundingRect.bottom) {
             return false;
         }
         viewportRect.offset(0, viewportRect.height() - pageRepeat);
@@ -348,7 +348,7 @@ public class PageManager {
     }
 
     public boolean prevViewport() {
-        if (viewportRect.top <= pagesBoundingRect.top) {
+        if ((int)viewportRect.top <= (int)pagesBoundingRect.top) {
             return false;
         }
         viewportRect.offset(0, -viewportRect.height() + pageRepeat);
